+++ workspaces/blackduck/plugins/blackduck-node/src/BlackDuckRestApi.ts
@@ -49,7 +49,7 @@ export class BlackDuckRestApi {
         },
       });
       const token = await auth.json();
-      this.logger.info('Auth Successfull');
+      this.logger.info('Auth Successful');
       this._bearer = token.bearerToken;
       return token.bearerToken;
     } catch (error) {
@@ -58,7 +58,7 @@ export class BlackDuckRestApi {
   }
   public async getProjects(name: string): Promise<BD_REST_API_RESPONSE> {
     const projects = await fetch(
-      `${this.host}/projects?limit=999&q=name:${name}`,
+      `${this.host}/projects?limit=999&q=${encodeURIComponent(`name:${name}`)}`,
       {
         method: 'GET',
         headers: {
@@ -68,7 +68,7 @@ export class BlackDuckRestApi {
         },
       },
     );
-    this.logger.debug('Retrived Projects!!');
+    this.logger.debug('Retrieved Projects!!');
     return projects.json();
   }
 
@@ -77,7 +77,9 @@ export class BlackDuckRestApi {
     versionName: string,
   ): Promise<BD_VERSIONS_API_RESPONSE> {
     const versions = await fetch(
-      `${projectUrl}/versions?limit=999&q=versionName:${versionName}`,
+      `${projectUrl}/versions?limit=999&q=${encodeURIComponent(
+        `versionName:${versionName}`,
+      )}`,
       {
         method: 'GET',
         headers: {
@@ -87,7 +89,7 @@ export class BlackDuckRestApi {
         },
       },
     );
-    this.logger.debug('Retrived Versions!!');
+    this.logger.debug('Retrieved Versions!!');
     return versions.json();
   }