+++ workspaces/entity-feedback/plugins/entity-feedback-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-entity-feedback-common
 
+## 0.4.0
+
+### Minor Changes
+
+- a5737fc: Backstage version bump to v1.36.1
+
 ## 0.3.0
 
 ### Minor Changes