+++ workspaces/scaffolder-backend-module-regex/plugins/regex-actions/CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Dependencies
 
+## 2.5.0
+
+### Minor Changes
+
+- 461e7fb: Backstage version bump to v1.36.1
+
 ## 2.4.0
 
 ### Minor Changes